requirements.txt
# Core AI & LLM Tools
openai
langchain
faiss-cpu
transformers
diffusers
accelerate
#xformers

# Deep Learning + GPU
torch
torchvision
torchaudio
GPUtil

streamlit
requests
pillow
numpy
langchain
openai
torch

# Backend & Frontend
fastapi
uvicorn
streamlit

# Image Handling
pillow

# 📊 Logging/Env
python-dotenv

# Optional (for experimenting/logging)
scipy
matplotlib


langchain-community
langchain-core
langchain-openai